✨ shared ui 내부 컴포넌트 shared/components로 분리
