+++ src/pages/postManager/ui/PostManager.tsx
@@ -3,30 +3,13 @@ import { useEffect, useState } from 'react';
 import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from 'lucide-react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
-import {
-  Button,
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  Input,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-  Textarea,
-} from '@/shared/ui';
+import { Card, CardContent, CardHeader, CardTitle } from '@/shared/components';
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/shared/components';
+import { Button } from '@/shared/components';
+import { Textarea } from '@/shared/components';
+import { Input } from '@/shared/components';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/shared/components';
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/shared/components';
 
 export const PostsManager = () => {
   const navigate = useNavigate();
