+++ src/shared/components/button/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './Button';
