+++ src/shared/components/card/ui/CardContent.tsx
@@ -1,0 +1,15 @@
+import { forwardRef, HTMLAttributes } from 'react';
+
+interface CardContentProps extends HTMLAttributes<HTMLDivElement> {}
+
+export const CardContent = forwardRef<HTMLDivElement, CardContentProps>(
+  ({ className, ...props }, ref) => (
+    <div
+      ref={ref}
+      className={`p-6 pt-0 ${className}`}
+      {...props}
+    />
+  ),
+);
+
+CardContent.displayName = 'CardContent';
