+++ src/shared/components/card/ui/index.ts
@@ -1,0 +1,4 @@
+export * from './Card';
+export * from './CardHeader';
+export * from './CardTitle';
+export * from './CardContent';
