+++ src/shared/components/dialog/ui/DialogHeader.tsx
@@ -1,0 +1,12 @@
+import { HTMLAttributes } from 'react';
+
+interface DialogHeaderProps extends HTMLAttributes<HTMLDivElement> {}
+
+export const DialogHeader = ({ className, ...props }: DialogHeaderProps) => (
+  <div
+    className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`}
+    {...props}
+  />
+);
+
+DialogHeader.displayName = 'DialogHeader';
