+++ src/shared/components/dialog/ui/index.ts
@@ -1,0 +1,4 @@
+export * from './DialogPrimitive';
+export * from './DialogContent';
+export * from './DialogHeader';
+export * from './DialogTitle';
