+++ src/shared/components/index.ts
@@ -1,0 +1,7 @@
+export * from './button';
+export * from './card';
+export * from './table';
+export * from './textarea';
+export * from './input';
+export * from './select';
+export * from './dialog';
