+++ src/shared/components/select/ui/index.ts
@@ -1,0 +1,4 @@
+export * from './SelectPrimitive';
+export * from './SelectContent';
+export * from './SelectItem';
+export * from './SelectTrigger';
