+++ src/shared/components/table/ui/TableBody.tsx
@@ -1,0 +1,15 @@
+import { forwardRef, HTMLAttributes } from 'react';
+
+interface TableBodyProps extends HTMLAttributes<HTMLTableSectionElement> {}
+
+export const TableBody = forwardRef<HTMLTableSectionElement, TableBodyProps>(
+  ({ className, ...props }, ref) => (
+    <tbody
+      ref={ref}
+      className={`[&_tr:last-child]:border-0 ${className}`}
+      {...props}
+    />
+  ),
+);
+
+TableBody.displayName = 'TableBody';
