+++ src/shared/components/table/ui/TableCell.tsx
@@ -1,0 +1,15 @@
+import { forwardRef, HTMLAttributes } from 'react';
+
+interface TableCellProps extends HTMLAttributes<HTMLTableCellElement> {}
+
+export const TableCell = forwardRef<HTMLTableCellElement, TableCellProps>(
+  ({ className, ...props }, ref) => (
+    <td
+      ref={ref}
+      className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`}
+      {...props}
+    />
+  ),
+);
+
+TableCell.displayName = 'TableCell';
