+++ src/shared/components/table/ui/TableHeader.tsx
@@ -1,0 +1,15 @@
+import { forwardRef, HTMLAttributes } from 'react';
+
+interface TableHeaderProps extends HTMLAttributes<HTMLTableSectionElement> {}
+
+export const TableHeader = forwardRef<HTMLTableSectionElement, TableHeaderProps>(
+  ({ className, ...props }, ref) => (
+    <thead
+      ref={ref}
+      className={`[&_tr]:border-b ${className}`}
+      {...props}
+    />
+  ),
+);
+
+TableHeader.displayName = 'TableHeader';
