+++ src/shared/components/table/ui/TableRow.tsx
@@ -1,0 +1,15 @@
+import { forwardRef, HTMLAttributes } from 'react';
+
+interface TableRowProps extends HTMLAttributes<HTMLTableRowElement> {}
+
+export const TableRow = forwardRef<HTMLTableRowElement, TableRowProps>(
+  ({ className, ...props }, ref) => (
+    <tr
+      ref={ref}
+      className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+      {...props}
+    />
+  ),
+);
+
+TableRow.displayName = 'TableRow';
