+++ src/shared/components/table/ui/index.ts
@@ -1,0 +1,6 @@
+export * from './Table';
+export * from './TableBody';
+export * from './TableCell';
+export * from './TableHead';
+export * from './TableHeader';
+export * from './TableRow';
